+++ cfdtools/__init__.py
@@ -2,7 +2,7 @@
 import logging
 import sys
 
-__version__ = "0.5.4"
+__version__ = "0.6.0"
 
 level = 'info'
 if '--log' in sys.argv: